+++ tests/testthat/test_read_images.R
@@ -34,3 +34,15 @@ test_that('read_maps works', {
   expect_equal(names(maps), expected_names)
   expect_equivalent(purrr::map(maps, dim), rep(list(c(260, 348)), 4))
 })
+
+test_that('get_field_info works', {
+  skip_if_not_installed('phenoptrExamples')
+  path = system.file('extdata', 'samples',
+                     'Set4_1-6plex_[16142,55840]_component_data.tif',
+                     package='phenoptrExamples')
+  info = get_field_info(path)
+  expect_equal(info$image_size, c(1868, 1400))
+  expect_equal(info$microns_per_pixel, 0.4981226, tolerance=0.0000001)
+  expect_equal(info$field_size, c(930.4929, 697.3716), tolerance=0.0001)
+  expect_equal(info$location, c(15676.75, 55491.31), tolerance=0.01)
+})