+++ NEWS.md
@@ -1,3 +1,9 @@
+# phenoptr 0.1.1.9002
+
+- Add `pixels_per_micron="auto"` option to `read_cell_seg_data()`.
+- Supports tiled component files using the PerkinElmer
+  [fork of the tiff package](https://github.com/PerkinElmer/tiff).
+
 # phenoptr 0.1.1.9001
 
 - Faster count_touching_cells when EBImage 4.19.9 is available.